Use inheritance to remove duplicate code
